Issues Update - Nov 20, 2017
-Client side code should not construct image urls (#139) - Fixed
-Tinkbox - Show contacts with most recent activity on top. (#124) - Done
-Menu - Iphone 5&6, Menu is to sensitive, pops up and down. (#133) -
Fixed
-Contact list - (cosmetic) Names that cannot be shown in one row is
broken in the middle. Can we make words not brake up in two rows? (#131)
- Cancelled by Nikhil
-Tinkfeed (cosmetic) - Name is broken in the middle, see attachment
(#130) - Cancelled by Nikhil, Same as the above one
-Contact list Contact picture is not shown in contact list. (Only
Nikhils picture) (#123) - We are able to pictures in the contacts list
of all testing accounts.
-Loading tink animation The old tinktime background (blue blurry) shows
before animation is loaded. Change to Tinktime background. (#125) -
Fixed - New background updated
-Tinkbox – smaller circle should be shown corresponding to either in or
out depending on balance. Now bigger circle is always on the left.
(#126) - Fixed